// src/app/models/asset-returning-event.ts
export interface AssetReturningEvent {
    eventId: string;
    assetId: string;
    assetName: string;
    requestOrgId: string;
    requestOrgIdentifier: string;
    requestOrgNam: string;
    org2: string;
    targetOrgId: string;
    targetOrgIdentifier: string;
    targetOrgNam: string;
    aggregateId: string;
    status: string;
    message: string;
    dateTimeFromNow: string;
}